feat: read the main three tables from TACO
Create logic to read the:
1. MacroAndMicronutrientsTable
2. FattyAcidsTable
3. AminoAcidsTable

They all share the same abstract base class Table.
Correction functions can be defined to handle specific cases, such as the misread columns in MacroAndMicronutrientsTable.

Also, this commit creates the TACO class to serve as a namespace for the functions that download and save the document.